fix(knowledge): change search ui and fix search bug
